more circular dependencies cleaned up
